This is a work in progress.
Add LICENSE and set to restricted because it's a Java 5 binary from the FreeBSD foundation.

Set USE_GCPIO so that CURRENT works with it.

Fix EST/EDT bug by adding parameters to tzupdate utility

Update TZUPDATE utility to the latest.

Change javavmwrapper installation so it works with fake. (it has to be present first)


git-svn-id: http://svn.midnightbsd.org/svn/mports@2197 c872e0db-cc11-e311-ab8b-002590791c19