Initial import of awesome window manager.
awesome is a tiling window manager initially based on a dwm code
rewriting. It's extremely fast, small, dynamic and awesome.

Windows can be managed in several layouts: tiled and floating. Each
layout can be applied on the fly, optimizing the environment for the
application in use and the task performed.

Managing windows in tiled mode assures that no space will be waste on
your screen. No gaps, no overlap.


git-svn-id: http://svn.midnightbsd.org/svn/mports@5269 c872e0db-cc11-e311-ab8b-002590791c19