Pretty big refactor of Chroot.pm.
A reference chroot dir is now created, and one or more working chroot dirs.
When a working chroot dir is marked dirty or dead, cpdup is used to sync
that dir with the reference dir.  This is more rebust and faster than the
older methods of cleaning a chroot dir.

The groundwork is also now in place for supporting more than one worker
chroot dir.  This will be used when the -j flag is implmented on the slave
side.


git-svn-id: http://svn.midnightbsd.org/svn/mports@6134 c872e0db-cc11-e311-ab8b-002590791c19