Add support for wacom tablets such as the one I have :)
Due to some poor hacks to the "run_configure" shell script in this port, it's currently only limited to i386.  It might work on amd64 and that should get resolved when I have time to test it.

Note this port installs a kernel module and modifies your x.org config file (once the setup command of the rc.d script is run).  It is based on the linux driver from wacom so it's under GPLv2.


git-svn-id: http://svn.midnightbsd.org/svn/mports@7601 c872e0db-cc11-e311-ab8b-002590791c19