Add a bandaid to this port to make it build again.
We've got a problem in general with imake ports currently.  imake has been hacked to run on midnightbsd for a long time, but now that we're properly reporting the OSREL to the ports system and a few other things, it's thinking we're at freebsd < 1.  Thus it tries gnumalloc and xpg4 (a lib from the freebsd4 era).

I've fixed the gnumalloc problem with imake, but i haven't figured out how to deal with this xpg bug correctly.


git-svn-id: http://svn.midnightbsd.org/svn/mports@7977 c872e0db-cc11-e311-ab8b-002590791c19