What did intel call it?  Multiplexity?  Yeah...
Execute multiple make jobs in parallel within a single port.

This uses make -jX where X is the number of CPUs on ports that
have this feature enabled to make use of those fancy multicore CPUs.

By default ports are built as they always have been.

Use DISABLE_MAKE_JOBS to turn this off globally (make.conf)
FORCE_MAKE_JOBS turns it on for everything in make.conf except those ports
with explicit MAKE_JOBS_UNSAFE set.

Finally port maintainers will use MAKE_JOBS_SAFE to mark a port as "OK" with
this feature.

Ports can use make or gmake with this option.

This is similar to the feature introduced in FreeBSD ports last year; in fact it's based on their approach.

Since this is used at compile time, it should not have a bearing on our faking or other features.  It's local to the wrksrc only!


git-svn-id: http://svn.midnightbsd.org/svn/mports@7993 c872e0db-cc11-e311-ab8b-002590791c19